♻️(lti) rework title edition and move live recording toggle
Make title editable onclick
Add edition icon on the right of the title
Move the toggle to record in Tools and application widget
