+++ src/frontend/apps/lti_site/components/StudentLiveInfoBar/index.spec.tsx
@@ -5,20 +5,40 @@ import { imageSnapshot } from 'utils/tests/imageSnapshot';
 import render from 'utils/tests/render';
 
 import { StudentLiveInfoBar } from '.';
+import { videoMockFactory } from 'lib-components';
+import { wrapInVideo } from '../../utils/tests/wrapInVideo';
 
 describe('<StudentLiveInfoBar />', () => {
   it('renders live title', () => {
-    render(<StudentLiveInfoBar title="title" startDate={null} />);
+    const mockedVideo = videoMockFactory({
+      allow_recording: false,
+      title: 'title',
+    });
 
-    screen.getByText('title');
+    render(wrapInVideo(<StudentLiveInfoBar startDate={null} />, mockedVideo), {
+      intlOptions: { locale: 'fr' },
+    });
+
+    screen.getByRole('heading', { name: 'title' });
   });
 
   it('renders live title and live start datetime [screenshot]', async () => {
+    const mockedVideo = videoMockFactory({
+      allow_recording: false,
+      title: 'title',
+    });
+
     render(
-      <StudentLiveInfoBar title={'title'} startDate={'2022-09-26T07:00:00Z'} />,
+      wrapInVideo(
+        <StudentLiveInfoBar startDate={'2022-09-26T07:00:00Z'} />,
+        mockedVideo,
+      ),
+      {
+        intlOptions: { locale: 'en' },
+      },
     );
 
-    screen.getByText('title');
+    screen.getByRole('heading', { name: 'title' });
     screen.getByText('9/26/2022  ·  7:00:00 AM', {
       normalizer: getDefaultNormalizer({ collapseWhitespace: false }),
     });
