+++ src/frontend/apps/lti_site/components/TeacherLiveInfoBar/index.spec.tsx
@@ -1,28 +1,30 @@
 import { getDefaultNormalizer, screen } from '@testing-library/react';
+import { videoMockFactory } from 'lib-components';
 import React from 'react';
 
 import render from 'utils/tests/render';
+import { wrapInVideo } from 'utils/tests/wrapInVideo';
 
 import { TeacherLiveInfoBar } from '.';
 
 describe('<TeacherLiveInfoBar />', () => {
   it('renders title and startDate', () => {
-    render(
-      <TeacherLiveInfoBar title={'title'} startDate={'2022-09-26T07:00:00Z'} />,
-    );
-
-    screen.getByRole('heading', { name: 'title' });
-    screen.getByText('9/26/2022  ·  7:00:00 AM', {
-      normalizer: getDefaultNormalizer({ collapseWhitespace: false }),
+    const mockedVideo = videoMockFactory({
+      allow_recording: false,
+      title: undefined,
     });
-  });
 
-  it('renders placeholder title when this one is missing', () => {
     render(
-      <TeacherLiveInfoBar title={null} startDate={'2022-09-26T07:00:00Z'} />,
+      wrapInVideo(
+        <TeacherLiveInfoBar startDate={'2022-09-26T07:00:00Z'} />,
+        mockedVideo,
+      ),
+      {
+        intlOptions: { locale: 'en' },
+      },
     );
 
-    screen.getByRole('heading', { name: 'No title' });
+    screen.getByDisplayValue('No title');
     screen.getByText('9/26/2022  ·  7:00:00 AM', {
       normalizer: getDefaultNormalizer({ collapseWhitespace: false }),
     });
